json updated filesize increased to long instead of int

+++ Models/Asset.cs
@@ -5,7 +5,7 @@
         public string url { get; set; }
         public int width { get; set; }
         public int height { get; set; }
-        public int fileSize { get; set; }
+        public long fileSize { get; set; }
         public string contentType { get; set; }
         public string type { get; set; }
     }
